pull out filters and comment
